Updated Assessments
Updated the Asessments portion of the site for the Teachers. Now each
assessment is linked to a teacher. No other teacher should be able to
manipulate it. Also updated the assessments table to allow this.

Minor UI tweaks elsewhere.

Additionally, part of asp.net core was updated. project.json was altered
to include a runtimes.