Applied Access Control
Applied some access control the the Assessments and Submissions
controllers.
Simple check when a User goes to edit something. If they don't own it,
they can't get to the edit page for it.

Additionally, the Submissions Create and Index pages were slightly
updated.
Index page removed the create link. Have to access the page from a link
on an Assessment.
Create page, removed the select box for an Assessment.